Changes implemented for final submission.

+++ src/main/java/org/ac/cst8277/kirk/patrick/usermanagementservice/controller/RoleController.java
@@ -4,10 +4,13 @@ import org.ac.cst8277.kirk.patrick.usermanagementservice.model.Response;
 import org.ac.cst8277.kirk.patrick.usermanagementservice.dao.MySQLUserDatabase;
 import org.ac.cst8277.kirk.patrick.usermanagementservice.dao.UserDatabase;
 import org.ac.cst8277.kirk.patrick.usermanagementservice.model.Role;
+import org.ac.cst8277.kirk.patrick.usermanagementservice.model.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.UUID;
 
 @RestController
 public class RoleController {
@@ -15,15 +18,23 @@ public class RoleController {
     public ResponseEntity<Response> createRole(@RequestBody Role role) {
         Response response = new Response();
 
-        role.generateId();
-
         UserDatabase database = new MySQLUserDatabase();
         database.open();
 
-        database.createRole(role);
-        response.setHttpStatus(HttpStatus.OK);
-        response.setMessage("Success");
-        response.setData(role);
+        User user = database.getUserByToken(role.getToken());
+
+        if (user != null) {
+            role.generateId();
+
+            database.createRole(role);
+            response.setHttpStatus(HttpStatus.OK);
+            response.setMessage("Success");
+            response.setData(role);
+        }
+        else {
+            response.setHttpStatus(HttpStatus.UNAUTHORIZED);
+            response.setMessage("User is not authenticated.");
+        }
 
         database.close();
 
@@ -37,26 +48,42 @@ public class RoleController {
         UserDatabase database = new MySQLUserDatabase();
         database.open();
 
-        database.updateRole(role);
-        response.setHttpStatus(HttpStatus.OK);
-        response.setMessage("Success");
-        response.setData(role);
+        User user = database.getUserByToken(role.getToken());
+
+        if (user != null) {
+            database.updateRole(role);
+            response.setHttpStatus(HttpStatus.OK);
+            response.setMessage("Success");
+            response.setData(role);
+        }
+        else {
+            response.setHttpStatus(HttpStatus.UNAUTHORIZED);
+            response.setMessage("User is not authenticated.");
+        }
 
         database.close();
 
         return new ResponseEntity<>(response, response.getHttpStatus());
     }
 
-    @DeleteMapping(value = "/role")
-    public ResponseEntity<Response> deleteRole(@RequestParam String name) {
+    @DeleteMapping(value = "/role/{name}")
+    public ResponseEntity<Response> deleteRole(@PathVariable String name, @RequestParam String token) {
         Response response = new Response();
 
         UserDatabase database = new MySQLUserDatabase();
         database.open();
 
-        database.deleteRole(name);
-        response.setHttpStatus(HttpStatus.OK);
-        response.setMessage("Success");
+        User user = database.getUserByToken(UUID.fromString(token));
+
+        if (user != null) {
+            database.deleteRole(name);
+            response.setHttpStatus(HttpStatus.OK);
+            response.setMessage("Success");
+        }
+        else {
+            response.setHttpStatus(HttpStatus.UNAUTHORIZED);
+            response.setMessage("User is not authenticated.");
+        }
 
         database.close();
